Oceans within BiomeGroups exterminated! (Hacky)
- Placed some temp fix code in `Layer` which removes null biome entries.
This was creating the situation where ocean was spawning in BiomeGroups
that didn't have ocean listed. Pretty sure the root cause is in my
handling of biome rarity
- Organized `Layer` bits a little
- Got BiomeMode setting working again
- Cleaned up `ReadBiomeGroups()` quite a bit!

TODO:
- Make a biome group with Ocean and Deep_Ocean for 1.7.2
- Make a biome group with Ocean for pre-1.7.2
- Add BiomeGroup rarity setting to replace LandRarity and IceRarity
- Add BiomeGroup size setting to replace LandSize and IceSize
- Add BiomeGroup temperature, determined upon load time as the average
of all biomes in the group
- Add WorldConfig setting for allowed variance of BiomeGroup
temperatures for 1.7-like biome differentiation (Ex. Snowy BiomeGroup
doesn't spawn next to DryWarm BiomeGroup unless you want it to)
- Fix Shores
- Fix Frozen Oceans
- Find a better way to handle Frozen Ocean with biomegroups...
Temperature based or BiomeConfig based? Not sure... I like the idea of
temperature based... But do we use individual BiomeConfig temp or the
BiomeGroup Average? Or both with a flag setting? ... Discussion needed.